requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-location-field==2.1.0
django-redis==4.10.0
django-rest-swagger==2.1.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
google-auth==1.6.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.4
openapi-codec==1.3.2
pandas==0.24.1
Pillow==5.4.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyJWT==1.7.1
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2018.9
querystring-parser==1.2.3
redis==3.0.1
requests==2.21.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.8
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
